+++ example/lib/main.dart
@@ -35,7 +35,7 @@ class _MyHomePageState extends State<MyHomePage> {
   final CurrencyTextFieldController _controller2 = CurrencyTextFieldController(
       initDoubleValue: -10, currencySeparator: ' -> ');
   final CurrencyTextFieldController _controller3 =
-      CurrencyTextFieldController(initIntValue: -1000, enableNegative: false);
+      CurrencyTextFieldController(initIntValue: -1000, enableNegative: false, maxValue: 2000);
 
   @override
   Widget build(BuildContext context) {